+++ cookiebanner/ext_emconf.php
@@ -1,7 +1,7 @@
 <?php
 $EM_CONF[$_EXTKEY] = [
     'title' => 'Cookiebanner',
-    'description' => 'Cookieconsent-Banner for DSGVO/GDPR',
+    'description' => 'Cookieconsent-Banner for DSGVO/GDPR! Cookie notice for your website.',
     'category' => 'plugin',
     'author' => 'Julian Seidl',
     'author_email' => 'hi@jseidl.at',
@@ -10,10 +10,10 @@ $EM_CONF[$_EXTKEY] = [
     'uploadfolder' => '0',
     'createDirs' => '',
     'clearCacheOnLoad' => 1,
-    'version' => '1.2.3',
+    'version' => '1.2.4-rc',
     'constraints' => [
         'depends' => [
-            'typo3' => '8.7.0-9.9.99',
+            'typo3' => '8.7.0-9.5.99',
         ],
         'conflicts' => [],
         'suggests' => [],